feat/ friendl play draft ok #104

+++ backend/nest_project/src/game/friendlyGame.gateway.ts
@@ -49,6 +49,7 @@ export class FriendlyGameGateway implements OnGatewayConnection, OnGatewayDiscon
 	async onModuleInit() 
 	{
 		this.connectedFriendlyPlayerService.deleteAll();
+		this.matchService.deleteAll();
 	}
 
 	async handleConnection(socket: Socket)
@@ -63,29 +64,37 @@ export class FriendlyGameGateway implements OnGatewayConnection, OnGatewayDiscon
 			return socket.disconnect();
 			// 클라이언트에서 이벤트 인지하면 menu로 리다이렉션
 		}
+		socket.data.userId = userId;
 		const userProfile = await this.profileService.getUserProfileById(userId);
 		if (!userProfile)
 		{
 			return socket.disconnect();
 			// 클라이언트에서 이벤트 인지하면 menu로 리다이렉션
 		}
+
 		const current = await this.connectedFriendlyPlayerService.createPlayer(userId, socket.id);
-		socket.emit('savePlayer', (inviteData) => {
-			this.connectedFriendlyPlayerService.updateInvitation(current, inviteData);
+		if(!current)
+		{
+			return socket.disconnect();
+		}
+		socket.emit('savePlayer', async (inviteData) => {
+			await this.connectedFriendlyPlayerService.updateInvitation(current, inviteData);
 		});
-		this.logger.log(`current Player: ${current.id}, ${current.socketId}`);
-		
-		if (this.connectedFriendlyPlayerService.isHostPlayer(current))
+		this.profileService.ingame(userId);
+		this.logger.log(`current Player : ${current.id}, ${current.socketId}`);
+
+		if (await this.connectedFriendlyPlayerService.isHostPlayer(current))
 		{
 			this.waitGame(current);
 		}
-		else if (this.connectedFriendlyPlayerService.isGuestPlayer(current))
+		else if (await this.connectedFriendlyPlayerService.isGuestPlayer(current))
 		{
 			this.acceptGame(current);
 		}
 		else
 		{
-			return this.endPlayer(socket.id, current.id);
+			this.profileService.logOn(socket.data.userId);
+			return socket.disconnect();
 		}
 	}
 
@@ -93,25 +102,60 @@ export class FriendlyGameGateway implements OnGatewayConnection, OnGatewayDiscon
 	{
 		// socket.emit('Error', new UnauthorizedException());
 
-		const player = (await this.connectedFriendlyPlayerService.getPlayerBySocketId(socket.id));
-		const match_id = (await this.matchService.getByPlayerId(player.id)).match_id;
-		if (match_id)
+		if (await this.connectedFriendlyPlayerService.getPlayerBySocketId(socket.id))
 		{
-			await this.endGame(match_id, player.id);
+			const player_id = (await this.connectedFriendlyPlayerService.getPlayerBySocketId(socket.id)).id;
+
+			if (await this.matchService.getByPlayerId(player_id) === null)
+			{
+				await this.connectedFriendlyPlayerService.deletePlayer(player_id);
+				this.profileService.logOn(socket.data.userId);
+				socket.disconnect();
+				return ;
+			}
+			else
+			{
+				const match_id = (await this.matchService.getByPlayerId(player_id)).match_id;
+				// this.logger.log(`handleDIsconnect : match id ${match_id}`);
+				if (match_id)
+				{
+					await this.endGame(match_id, player_id);
+				}
+				await this.connectedFriendlyPlayerService.deletePlayer(player_id);
+				this.profileService.logOn(socket.data.userId);
+				socket.disconnect();
+				return ;
+			}
 		}
+
+		if ((await this.connectedFriendlyPlayerService.getPlayerBySocketId(socket.id)) === null)
+		{
+			this.profileService.logOn(socket.data.userId);
+			socket.disconnect();
+			return ;
+		}
+		const player = (await this.connectedFriendlyPlayerService.getPlayerBySocketId(socket.id));
+		// const match_id = (await this.matchService.getByPlayerId(player.id)).match_id;
+		// if (match_id)
+		// {
+		// 	await this.endGame(match_id, player.id);
+		// }
 
 		if (player.id === player.hostId)
 		{
-			clearInterval(player.checkTimer);
+			await clearInterval(player.checkTimer);
 		}
 
-		this.endPlayer(socket.id, player.id);
+		this.connectedFriendlyPlayerService.deletePlayer(player.id);
+		this.profileService.logOn(socket.data.userId);
+		socket.disconnect();
+		// this.endPlayer(socket.id, player.id);
 	}
 
 	private async endPlayer(socket_id: string, player_id: number)
 	{
 		// socket.emit('Error', new UnauthorizedException());
-		this.connectedFriendlyPlayerService.deletePlayer(player_id);
+		// this.connectedFriendlyPlayerService.deletePlayer(player_id);
 		this.server.in(socket_id).disconnectSockets(true);
 		this.logger.log(`Friendly Game Server: socketId [ ${socket_id} ] disconnected.`);
 	};
@@ -134,7 +178,7 @@ export class FriendlyGameGateway implements OnGatewayConnection, OnGatewayDiscon
 	private async waitGame(host: FriendlyPlayer)
 	{
 		const currentTime = Date.now();
-		host.checkTimer = setInterval(this.checkRefuse, 1000, host, currentTime) as unknown as number;
+		host.checkTimer = setInterval(() => this.checkRefuse(host, currentTime), 1000) as unknown as number;
 	}
 
 	private async acceptGame(guest: FriendlyPlayer)
@@ -147,6 +191,11 @@ export class FriendlyGameGateway implements OnGatewayConnection, OnGatewayDiscon
 	{
 		const host = await this.connectedFriendlyPlayerService.getHostbySocketId(socket.id);
 		const guest = await this.connectedFriendlyPlayerService.getPlayerBySocketId(socket.id);
+		if (!host)
+		{
+			this.endPlayer(socket.id, guest.id);
+			return ;
+		}
 		const currentMatch = await this.matchService.createCustom(host.id, guest.id, gameType);
 		if (!currentMatch)
 		{
@@ -154,6 +203,7 @@ export class FriendlyGameGateway implements OnGatewayConnection, OnGatewayDiscon
 			return ;
 		}
 		clearInterval(host.checkTimer);
+		this.logger.log(`setGame : match ${currentMatch.match_id} will soon start!`);
 		this.startGame(currentMatch);
 	}
 
@@ -183,16 +233,16 @@ export class FriendlyGameGateway implements OnGatewayConnection, OnGatewayDiscon
 			ballX: 800 / 2,
 			ballY: 600 / 2,
 			ballRadius: 10,
-			ballXvelocity: 3,
-			ballYvelocity: 3,
-			ballSpeed: 3,
+			ballXvelocity: 6,
+			ballYvelocity: 6,
+			ballSpeed: 10,
 			matchId: match.match_id,
 			gameTimer: null,
 		}
 
 		if (match.game_type === "speedUp")
 		{
-			gameField.ballSpeed = 6;
+			gameField.ballSpeed = 14;
 		}
 		else if (match.game_type === "smallBall")
 		{
@@ -200,13 +250,13 @@ export class FriendlyGameGateway implements OnGatewayConnection, OnGatewayDiscon
 		}
 		else if (match.game_type === "enjoyAll")
 		{
-			gameField.ballSpeed = 6;
+			gameField.ballSpeed = 14;
 			gameField.ballRadius = 5;
 		}
 
 		this.logger.log(`friendly/startGame : ${match.match_id} -> ${player1.id} vs ${player2.id}`);
 		this.gameFieldArr.push(gameField);
-		gameField.gameTimer = setInterval(playGame, 30, this.server, match, player1, player2, gameField);
+		gameField.gameTimer = setInterval(() => { this.playGame(this.server, match, player1, player2, gameField); }, 20);
 	}
 
 	private async getGameFieldByMatchId(match_id: number)
@@ -214,7 +264,7 @@ export class FriendlyGameGateway implements OnGatewayConnection, OnGatewayDiscon
 		for (let i = 0; i < this.gameFieldArr.length; ++i)
 		{
 			if (this.gameFieldArr[i].matchId === match_id)
-			return this.gameFieldArr[i];
+				return this.gameFieldArr[i];
 		}
 	}
 
@@ -222,6 +272,8 @@ export class FriendlyGameGateway implements OnGatewayConnection, OnGatewayDiscon
 	async movePlayer(@ConnectedSocket() socket: Socket, @MessageBody() userY: number)
 	{
 		const player = await this.connectedFriendlyPlayerService.getPlayerBySocketId(socket.id);
+		if (!player)
+			return ;
 		const match = (await this.matchService.getByPlayerId(player.id));
 		const opponent = await this.matchService.getOpponentByPlayerId(match.match_id, player.id);
 		const gameField = await this.getGameFieldByMatchId(match.match_id);
@@ -235,17 +287,22 @@ export class FriendlyGameGateway implements OnGatewayConnection, OnGatewayDiscon
 			userY = 500;
 		}
 
+		const data = {
+			leftY: gameField.paddleLeftY,
+			rightY: gameField.paddleRightY,
+		};
+
 		if (match.playerLeft === player.id)
 		{
 			gameField.paddleLeftY = userY;
-			this.server.to(player.socketId).emit('paddleMove', gameField);
-			this.server.to(opponent.socketId).emit('paddleMove', gameField);
+			this.server.to(player.socketId).emit('paddleMove', data);
+			this.server.to(opponent.socketId).emit('paddleMove', data);
 		}
 		else if (match.playerRight === player.id)
 		{
 			gameField.paddleRightY = userY;
-			this.server.to(player.socketId).emit('paddleMove', gameField);
-			this.server.to(opponent.socketId).emit('paddleMove', gameField);
+			this.server.to(player.socketId).emit('paddleMove', data);
+			this.server.to(opponent.socketId).emit('paddleMove', data);
 		}
 	}
 
@@ -262,10 +319,10 @@ export class FriendlyGameGateway implements OnGatewayConnection, OnGatewayDiscon
 
 	async endGame(match_id: number, loser_id: number)
 	{
-		this.logger.log(`endGame : match finished.`);
+		this.logger.log(`endGame : ${match_id} match finished.`);
 
 		const gameField = await this.getGameFieldByMatchId(match_id);
-		clearInterval(gameField.gameTimer);
+		await clearInterval(gameField.gameTimer);
 		const match = await this.matchService.getByMatchId(match_id);
 		let winner_id = 0;
 
@@ -303,45 +360,8 @@ export class FriendlyGameGateway implements OnGatewayConnection, OnGatewayDiscon
 		this.endPlayer((await this.connectedFriendlyPlayerService.getPlayer(loser_id)).socketId, loser_id);
 	}
 
-}
 
-
-async function collision(b: Ball, p: Paddle)
-{
-	const paddleLocation =
-	{
-	  top: p.y,
-	  bottom: p.y + p.height,
-	  left: p.x,
-	  right: p.x + p.width,
-	};
-
-	const ballLocation =
-	{
-	  top: b.y - b.radius,
-	  bottom: b.y + b.radius,
-	  left: b.x - b.radius,
-	  right: b.x + b.radius,
-	};
-
-	return (
-	  ballLocation.right > paddleLocation.left &&
-	  ballLocation.left < paddleLocation.right &&
-	  ballLocation.top < paddleLocation.bottom &&
-	  ballLocation.bottom > paddleLocation.top
-	);
-  }
-
-async function resetBall(gameField: GameField)
-{
-	gameField.ballX = gameField.canvasWidth / 2;
-	gameField.ballY = gameField.canvasHeight / 2;
-	gameField.ballXvelocity = -3;
-	gameField.ballYvelocity = 3;
-	gameField.ballSpeed = 3;
-}
-
-async function playGame(server: Server, match: MatchEntity, player1: Player, player2: Player, gameField: GameField)
+	async playGame(server: Server, match: MatchEntity, player1: Player, player2: Player, gameField: GameField)
 {
 	// location of ball
 	gameField.ballX += gameField.ballXvelocity;
@@ -396,26 +416,81 @@ async function playGame(server: Server, match: MatchEntity, player1: Player, pla
 	if (gameField.ballX - gameField.ballRadius < 0)
 	{
 		++gameField.scoreRight;
-		this.MatchService.updateRightScore(match.match_id, gameField.scoreRight);
+		this.matchService.updateRightScore(match.match_id, gameField.scoreRight);
 
-		if (gameField.scoreRight === 7)
+		if (gameField.scoreRight > 2)
 		{
-			this.endGame(match, null);
+			this.endGame(match.match_id, null);
 		}
-		resetBall(gameField);
+		if (match.game_type === "speedUp" || match.game_type === "enjoyAll")
+			resetBall(gameField, 14, -1);
+		else
+			resetBall(gameField, 10, -1);
 	}
 	else if (gameField.ballX + gameField.ballRadius > gameField.canvasWidth)
 	{
 		++gameField.scoreLeft;
-		this.MatchService.updateLeftScore(match.match_id, gameField.scoreRight);
+		this.matchService.updateLeftScore(match.match_id, gameField.scoreRight);
 
-		if (gameField.scoreLeft === 7)
+		if (gameField.scoreLeft > 2)
 		{
-			this.endGame(match, null);
+			this.endGame(match.match_id, null);
 		}
-		resetBall(gameField);
+		if (match.game_type === "speedUp" || match.game_type === "enjoyAll")
+			resetBall(gameField, 14, 1);
+		else
+			resetBall(gameField, 10, 1);
 	}
 
-	server.to(player1.socketId).emit('updateCanvas', gameField);
-	server.to(player2.socketId).emit('updateCanvas', gameField);
+	const data = {
+		ballX: gameField.ballX,
+		ballY: gameField.ballY,
+		veloX: gameField.ballXvelocity,
+		veloY: gameField.ballYvelocity,
+		ballSpeed: gameField.ballSpeed,
+		leftScore: gameField.scoreLeft,
+		rightScore: gameField.scoreRight,
+	};
+
+	server.to(player1.socketId).emit('updateCanvas', data);
+	server.to(player2.socketId).emit('updateCanvas', data);
+	}
+
 }
+
+
+async function collision(b: Ball, p: Paddle)
+{
+	const paddleLocation =
+	{
+	  top: p.y,
+	  bottom: p.y + p.height,
+	  left: p.x,
+	  right: p.x + p.width,
+	};
+
+	const ballLocation =
+	{
+	  top: b.y - b.radius,
+	  bottom: b.y + b.radius,
+	  left: b.x - b.radius,
+	  right: b.x + b.radius,
+	};
+
+	return (
+	  ballLocation.right > paddleLocation.left &&
+	  ballLocation.left < paddleLocation.right &&
+	  ballLocation.top < paddleLocation.bottom &&
+	  ballLocation.bottom > paddleLocation.top
+	);
+  }
+
+async function resetBall(gameField: GameField, speed: number, direction: number)
+{
+	gameField.ballX = gameField.canvasWidth / 2;
+	gameField.ballY = gameField.canvasHeight / 2;
+	gameField.ballXvelocity = 6 * direction;
+	gameField.ballYvelocity = 6;
+	gameField.ballSpeed = speed;
+}
+
